♻️ refactor(constants): update timing window type to double
- change TimmingWindowMs dictionary value type to double
- remove float suffixes for consistency with double type

+++ Winithm.Core/Scripts/Constants/HitResult.cs
@@ -5,13 +5,13 @@ namespace Winithm.Core.Constants
 {
   public static class HitResult
   {
-    public static readonly Dictionary<HitResultType, float> TimmingWindowMs = new Dictionary<HitResultType, float>
+    public static readonly Dictionary<HitResultType, double> TimmingWindowMs = new Dictionary<HitResultType, double>
     {
-      { HitResultType.Perfect, 65f },
-      { HitResultType.Good, 125f },
-      { HitResultType.Bad, 175f },
+      { HitResultType.Perfect, 65 },
+      { HitResultType.Good, 125 },
+      { HitResultType.Bad, 175 },
       // Miss is not a timing window, but a flag for notes that were not hit.
-      { HitResultType.Miss, 175f },
+      { HitResultType.Miss, 175 },
     };
 
     public static readonly Dictionary<HitResultType, float> ResultWeight = new Dictionary<HitResultType, float>
